Foi realizada a refatoração do código e melhoria do mesmo, além de criação de interfaces para melhorar o contrato

+++ shopping-api/src/main/java/com/adam/backend/shoppingapi/dtos/ProductDTO.java
@@ -17,7 +17,7 @@ public class ProductDTO {
     @NotBlank
     private String nome;
     @NotNull
-    private Float preco;
+    private Double preco;
 
     private String descricao;
